# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Pelican MC Router API
  version: 1.0.7
  description: Stable v1 management API. Credentials and raw backend diagnostics are never returned.
paths:
...
  /dashboard:
    get: {operationId: dashboard, security: [{oidcBearer: []}], responses: {'200': {description: Read-only operations dashboard, content: {text/html: {schema: {type: string}}}}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}, '503': {$ref: '#/components/responses/Error'}}}
...
  /api/v1/dashboard/reconcile:
    post: {operationId: reconcileDashboard, security: [{oidcBearer: []}], parameters: [{name: X-Pelican-MC-Router-CSRF, in: header, required: true, schema: {type: string, enum: ['1']}}], responses: {'200': {$ref: '#/components/responses/ManualReconcile'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '408': {$ref: '#/components/responses/Error'}, '429': {$ref: '#/components/responses/Error'}, '503': {$ref: '#/components/responses/Error'}}}
...
components:
  securitySchemes:
    bootstrapBearer: {type: http, scheme: bearer, description: Mounted one-time bootstrap token; only while setup is incomplete.}
    oidcBearer: {type: http, scheme: bearer, bearerFormat: JWT, description: Verified OIDC bearer. Viewer reads topology; operator performs mutations.}
...
  responses:
    Error: {description: Sanitized error, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    ManualReconcile: {description: Reconciliation completed, content: {application/json: {schema: {$ref: '#/components/schemas/ManualReconcile'}}}}
  schemas:
    Error: {type: object, required: [error], properties: {error: {type: string}}}
...
    RouteChanges: {type: object, required: [desired, created, updated, deleted, changed], properties: {desired: {type: integer}, created: {type: integer}, updated: {type: integer}, deleted: {type: integer}, changed: {type: boolean}}}
    Reconciliation: {type: object, properties: {in_progress: {type: boolean}, last_outcome: {type: [string, 'null'], enum: [not_configured, success, failure, null]}, last_started_at: {type: [string, 'null'], format: date-time}, last_completed_at: {type: [string, 'null'], format: date-time}, last_success_at: {type: [string, 'null'], format: date-time}, last_duration_ms: {type: integer}, consecutive_failures: {type: integer}, last_error: {type: [string, 'null']}, route_changes: {$ref: '#/components/schemas/RouteChanges'}}}
...
    ManualReconcile: {type: object, required: [reconciliation], properties: {reconciliation: {$ref: '#/components/schemas/Reconciliation'}}}